Work with images in memory bypassing tempfiles
I wasn't happy with writing image bin out to tempfiles to work with.
Instead I'm keeping images in memory, transforming in memory, and
writing back to the DB from memory. Now the only tempfile should be
during downloading the original image. Maybe I can get rid of that in the
future.